Remove Internal Test Code
At ViewController file, you can change the view displayed in ContentView to ListView, CollectionView, SignupView. The examples are not in English and may display unnecessary content.

Since this examples were developed using the current unpublished version of UICreator, it doesn't represent the final code. Some piece of code may be using some architecture of building like in SignupView that can be done different if I were using the lasted version.